Try to avoid using innerHTML or make it safe
innerHTML has XSS potential, it was used three times. Two could be
replaced by textContent and the last one that is still used was made
safe. The vulnerability could have been exploited if the user of the
addon would have been tricked to add attacker's malformed title or url.
Now, only dynamic things there are in the innerHTML string, are
translations and either 'checked' or ''. Translations don't pose a
threat.