Stay in a container
If browser supports containers, the user has enabled stay in a
container option, containers are enabled in the browser and the tab user
is about to redirect is in a container, open the redirected tab in the
same container.

Firefox and Firefox for Android are the only browsers that supports
containers. The support can be checked only vaguely, it can be checked
with contextualIdentities API, but adding contextualIdentities
permission enables containers every time. Some users don't use
containers, so it can't be used.
This is the reason open in a container option is always shown on
Firefox, because we can't know whether containers are supported and
enabled or not.