Add a browser action
Add a browser action for an alternative way to redirect a page, and only
page, you can't redirect a link with browser action.

A browser action can be useful when a page has disabled the context menu,
you can still open the context menu via Shift + right mouse button, but
users might not know about it or have forgotten it.
The other reason for having a browser action, is that there can be many
menus in the context menu and it can be difficult to find. Although,
there can be many browser action icons too. Maybe some users prefere
browser action over context menu.