Open new tab with a long press on android
Redirect to a new tab by pressing long on Firefox for Android. Right
click, in addition to middle click, opens a new tab also on desktop
browsers.

TODO

Chrome needs an option to always open a new tab, because it doesn't have
an API for context menu button press or modifiers.