Add an option for Chrome to open a new tab always
Chrome doesn't have an API for which button was pressed or even key
modifiers, that's why we add an option to always open a new tab for
a redirection. The browser action works in the same fashion as in
Firefox, left click redirecst current tab, middle or right opens a
new tab. In Firefox for Android normal touch redirects current tab
and long touch opens a new tab.