Add always open to new tab option
Add always open link or page to new tab option. Doesn't exist on Firefox for Android because it isn't supported (TODO check has this changed). On Firefox if this is off, you can open to new tab if you use middle mouse button. And on Chrome middle mouse button isn't supported at all (could be in browser action, but I decided it's more logical if it doesn't work in browser action either, because it doesn't work in context menus).